requirements.txt
Django>=4.0.2,<4.1.0
djangorestframework>=3.13.1,<3.14.0
psycopg2>=2.9.3,<2.10.0

flake8>=4.0.1,<4.1.0
